requirements.txt
dependency-injector==4.31.2
flask==1.1.2
bootstrap-flask==1.5.2
pyyaml==5.4.1
pygithub==1.54.1